Remove input and output tensor caching
Remove caches for input and output tensors to an executable. We use TF
pointer to create a nGraph tensor in which case there really isn't much
benefit in trying to cache input or output tensors.

Because the executor doesn't plug in as a device, it is the backend's
responsibility to copy between host and device memories as necessary.